Added new file client.go and client_test.go and moved relevant code there.
